+++ frontend/cypress/integration/common/istio_config.ts
@@ -1,43 +1,10 @@
 import { AfterAll, Given, Then, When } from '@badeball/cypress-cucumber-preprocessor';
 import { colExists, getColWithRowText } from './table';
-import { ensureKialiFinishedLoading } from './transition';
+import { ensureKialiFinishedLoading, waitForResourceDeletion } from './transition';
 import { getGVKTypeString } from 'utils/IstioConfigUtils';
 
 const CLUSTER1_CONTEXT = Cypress.env('CLUSTER1_CONTEXT');
 const CLUSTER2_CONTEXT = Cypress.env('CLUSTER2_CONTEXT');
-
-const waitForResourceDeletion = (
-  namespace: string,
-  kind: string,
-  name: string,
-  timeoutMs = 120000,
-  pollIntervalMs = 3000
-): Cypress.Chainable<unknown> => {
-  const startTime = Date.now();
-  const command = `kubectl -n ${namespace} get ${kind} ${name} --ignore-not-found -o name`;
-
-  const waitUntilDeleted = (): Cypress.Chainable<unknown> => {
-    return cy.exec(command, { failOnNonZeroExit: false }).then(result => {
-      const resourceName = result.stdout.trim();
-      if (resourceName === '') {
-        cy.log(`${kind}/${name} was deleted from namespace ${namespace}`);
-        return cy.wrap(null);
-      }
-
-      if (Date.now() - startTime >= timeoutMs) {
-        throw new Error(
-          `Timed out after ${timeoutMs}ms waiting for ${kind}/${name} to be deleted from namespace ${namespace}`
-        );
-      }
-
-      cy.log(`${kind}/${name} still exists in namespace ${namespace}; waiting ${pollIntervalMs}ms before retry`);
-      cy.wait(pollIntervalMs);
-      return waitUntilDeleted();
-    });
-  };
-
-  return waitUntilDeleted();
-};
 
 const labelsStringToJson = (labelsString: string): string => {
   let labelsJson = '';
